* consumer groups mapping moved to mapper class * cg page size moved to config * cg pagination test added
